+++ animations/ninja.json
@@ -1,108 +1,117 @@
 [
-    
     {
-        "spritesheet": "images\\ninja_walk_2.png",
-        "colorkey": [255, 255, 255],
-        "width": 37,
-        "height": 107,
-        "animations": {
-            "walk": {
+        "spritesheet": "images/ninja/default.png",
+        "width":49,
+        "height":78,
+        "colorkey":[0,0,0],
+        "scale":0.7,
+        "animations":{
+            "default":{
                 "loop": true,
+                "frames": 9,
+                "column":0,
+                "row":0,
+                "direction":"x"
+            }, 
+            "shoot":{
+                "loop": false,
                 "frames": 2,
-                "column": 0,
-                "row": 0,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_run2.png",
-        "colorkey": [255, 255, 255],
-        "width": 32,
-        "height": 107,
-        "animations": {
-            "run": {
+                "column":10,
+                "row":0,
+                "direction":"x"
+            }, 
+            "jump":{
                 "loop": true,
                 "frames": 1,
-                "column": 2,
-                "row": 0,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_fall2.png",
-        "colorkey": [255, 255, 255],
-        "width": 58,
-        "height": 107,
-        "animations": {
-            "squat": {
+                "column":11,
+                "row":0,
+                "direction":"x"
+            }, 
+            "fall":{
                 "loop": true,
                 "frames": 1,
-                "column": 0,
-                "row": 1,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_sword_attack2.png",
-        "colorkey": [255, 255, 255],
-        "width": 104,
-        "height": 107,
-        "animations": {
-            "ready_instance": {
-                "loop": true,
-                "frames": 1,
-                "column": 0,
-                "row": 2,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_highkick2.png",
-        "colorkey": [255, 255, 255],
-        "width": 89,
-        "height": 107,
-        "animations": {
-            "vanishing_step": {
-                "loop": true,
-                "frames": 1,
-                "column": 0,
-                "row": 3,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_special_attack2.png",
-        "colorkey": [255, 255, 255],
-        "width": 98,
-        "height": 107,
-        "animations": {
-            "fall": {
-                "loop": true,
-                "frames": 1,
-                "column": 3,
-                "row": 0,
-                "direction": "x"
-            }
-        }
-    },
-    {
-        "spritesheet": "images\\ninja_knockout2.png",
-        "colorkey": [255, 255, 255],
-        "width": 88,
-        "height": 107,
-        "animations": {
-            "sword_attack": {
+                "column":12,
+                "row":0,
+                "direction":"x"
+            },
+            "protect":{
                 "loop": false,
                 "frames": 1,
-                "column": 0,
-                "row": 1,
-                "direction": "x"
+                "column":9,
+                "row":0,
+                "direction":"x"
+            }
+        }
+    },
+    {
+        "spritesheet": "images/ninja/jump.png",
+        "width":31,
+        "height":107,
+        "colorkey":[0,0,0],
+        "scale":0.7,
+        "animations":{
+            "fallh":{
+                "loop": true,
+                "frames": 1,
+                "column":0,
+                "row":0,
+                "direction":"x"
+            }, 
+            "jumph":{
+                "loop": true,
+                "frames": 1,
+                "column":1,
+                "row":0,
+                "direction":"x"
+            }
+        }
+    },
+    {
+        "spritesheet": "images/ninja/punch.png",
+        "width":85,
+        "height":78,
+        "colorkey":[255,255,255],
+        "scale":0.7,
+        "animations":{
+            "punch":{
+                "loop": true,
+                "frames": 1,
+                "column":0,
+                "row":0,
+                "direction":"x"
+            }
+        }
+    },
+    {
+        "spritesheet": "images/ninja/die.png",
+        "width":49,
+        "height":78,
+        "colorkey":[0,0,0],
+        "scale":0.7,
+        "animations":{
+            "die":{
+                "loop": true,
+                "frames": 3,
+                "column":0,
+                "row":0,
+                "direction":"x"
+            }
+        }
+    },
+    {
+        "spritesheet": "images/ninja/walk.png",
+        "width":70,
+        "height":78,
+        "colorkey":[0,0,0],
+        "scale":0.7,
+        "animations":{
+            "walk":{
+                "loop": true,
+                "frames": 5,
+                "column":0,
+                "row":0,
+                "direction":"x"
             }
         }
     }
-]    
+]
